VFS-4808 Pass message string to lager as argument instead of format
Since message formatting is done on our side, lager needs not
perform any. However passing the resulting string alone causes tilde
characters in it to be treated as format specifiers and trigger
BAD FORMAT exception. When passing the formatted string as a format
argument it will not be parsed for format specifiers.